% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zurand-package.R
\docType{package}
\name{zurand-package}
\alias{zurand}
\alias{zurand-package}
\title{zurand: Stateless Random Numbers via Counter-Based Generation}
\description{
Stateless deterministic random number generation using the Philox counter-based generator from the 'Random123' library by Salmon, Moraes, Dror and Shaw (2011) \doi{10.1145/2063384.2063405}. Random values are pure functions of immutable keys and sampler arguments, so sampling never reads or mutates '.Random.seed'. Independent substreams are requested as key vectors with rng_key(seed, n) or derived explicitly with rng_fold().
}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item D. E. Shaw Research (Random123 library, src/Random123/) [copyright holder]
  \item NumPy Developers (Ziggurat constant tables, src/numpyzig/) [copyright holder]
  \item Sun Microsystems, Inc. (fdlibm exp() and log1p(), src/zurand_fdlibm.h) [copyright holder]
}

}
\keyword{internal}
